+++ packages/animated-pose/src/types.ts
@@ -31,6 +31,7 @@ export type Transition = (
 export type CreateValueProps = {
   passiveParent?: Value;
   passiveProps?: Interpolation;
+  props?: { [key: string]: any };
 };
 
 export type AnimatedPoseConfig = PoserConfig<Value>;