+++ packages/popmotion-react/.storybook/addons.js
@@ -1,4 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
-
-import '@storybook/addon-actions/register';
-import '@storybook/addon-links/register';